+++ Local/src/Dap.Local/EmbeddedResource.fs
@@ -1,6 +1,7 @@
 [<AutoOpen>]
 module Dap.Local.EmbeddedResource
 
+open System.IO;
 open System.Reflection;
 
 open Dap.Prelude
@@ -14,8 +15,21 @@ type EmbeddedResource = EmbeddedResourceHelper with
         let assembly = assembly |> Option.defaultValue (Assembly.GetCallingAssembly ())
         assembly.GetManifestResourceNames ()
     static member LogNames (logger : ILogger, ?assembly : Assembly) =
-        let assembly = assembly |> Option.defaultValue (Assembly.GetCallingAssembly ())
-        EmbeddedResource.GetNames (assembly = assembly)
+        EmbeddedResource.GetNames (?assembly = assembly)
         |> Array.iter (fun name ->
             logWarn logger "EmbeddedResource" "ResourceName" name
         )
+    static member TryCreateFromStream<'v> (relPath : string, create : (System.IO.Stream -> 'v), ?logger : ILogger, ?assembly : Assembly) =
+        let assembly = assembly |> Option.defaultValue (Assembly.GetCallingAssembly ())
+        let name = EmbeddedResource.GetName (relPath, assembly = assembly)
+        try
+            use stream = assembly.GetManifestResourceStream (name)
+            Some <| create stream
+        with e ->
+            let logger =
+                logger
+                |> Option.defaultWith (fun () -> getLogging () :> ILogger)
+            logException logger "EmbeddedResource.FromStream" (typeof<'v>.FullName) (name) e
+            None
+
+